Isolate S3FileKeySensor
