+++ airflow_home/plugins/operators/__init__.py
@@ -1,5 +1,5 @@
+from plugins.operators.s3_file_key_sensor import S3FileKeySensor
 from plugins.operators.s3_get_key_operator import S3GetKeyOperator
 from plugins.operators.s3_key_sensor import (
-    S3FileKeySensor,
     S3WildcardKeySensor,
 )
